requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1

# LLM Providers
openai==1.3.7
anthropic==0.7.7
google-generativeai==0.3.2
groq==0.4.2
cohere==4.37
ollama==0.1.7
asyncio-throttle==1.0.2

# WebSocket Support
websockets==12.0

# For Vercel deployment
requests==2.31.0

# Email validation for Pydantic
email-validator==2.1.0

# OAuth Support
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
PyJWT==2.8.0

# PDF Generation (optional) - commented out due to build issues
# reportlab==4.0.7
# Pillow==10.1.0

# Mistral AI - commented out due to build issues
# mistralai==0.0.12 